requirements.txt
Flask==2.1.3
python-dotenv==0.20.0
schedule==1.1.0
telebot==0.0.4
pyppeteer==1.0.2
